requirements.txt
Flask==1.1.2
humanize
requests
sendgrid
pyjwt

google-cloud-datastore<2.0.0dev
google-cloud-ndb
google-cloud-storage
google-cloud-vision


